Refactor: Footer 삭제,  NavBar로 통합 및 수정
